Split shard_spec/parse_shard and sink_names/validate_sink_name out of main.hpp
Neither depends on anything TTS_MAIN-specific, unlike the driver
function and report_pass/report_fail/report_fatal that stay - moves
them to their own focused files (engine/shard.hpp, engine/
sink_selection.hpp) instead of letting main.hpp keep absorbing every
new CLI flag's supporting logic.
